Move tests for database error logging
